Export csv files to other file formats as defined @param object is a character string of the csv file (with its path) or an R data.frame.

Export csv files to other file formats as defined @param object is a character string of the csv file (with its path) or an R data.frame.

dpr_export(
  object,
  export_format = NULL,
  export_folder,
  details = TRUE,
  readme = NULL
)

Arguments

export_format
is a character vector of formats to convert. Defaults to ".rds". Potential file types are .xlsx, .rds, .sav, .dta, .sas7bdat, and .json

export_folder
is the folder where the exported data will be placed. In this folder the a subfolder will be created for each csv file name and selected export_format file types will be created in the folder.

details
If TRUE then information is printed to screen during the process. Default is TRUE.

readme
is the data summary that describes the data. Accepts a text string describing the provenance.

Examples

dd <- read_csv(system.file("extdata", "Draft_vietnam.csv", package = "DataPushR"))

#> Error in read_csv(system.file("extdata", "Draft_vietnam.csv", package = "DataPushR")): could not find function "read_csv"

tempf <- fs::file_temp("Draft_vietnam_-_", ext = "csv")
write_csv(dd, tempf)

#> Error in write_csv(dd, tempf): could not find function "write_csv"

dpr_export(tempf, export_format = c(".rds", ".xlsx", ".sav", ".dta", ".csv", ".json", ".sas7bdat"),
export_folder=getwd())

#> Error in fs::path_file(csv_file) %>% fs::path_ext_remove(): could not find function "%>%"

dpr_export(dd, export_format=c(".rds",".xlsx",".sav",".dta",".csv", ".json", ".sas7bdat"), export_folder=getwd())

#> Error in fs::path_file(csv_file) %>% fs::path_ext_remove(): could not find function "%>%"